requirements.txt
streamlit
numpy
matplotlib
opencv-python-headless
Pillow
ultralytics
torch==2.0.0